backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.1.4
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
